% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/new_factor_function13.R
\name{add.factormean}
\alias{add.factormean}
\title{Calculates mean response leaving an observation out.}
\usage{
add.factormean(factor, response)
}
\arguments{
\item{factor}{Factor is the variable we want to calculate factor mean for.}

\item{response}{Response variable with values {0,1}.}
}
\description{
The function calculates mean response for each level of factor variable
while leaving out the current observation. It outputs mean response
for each observation. As the level depends on the output, it cannot be
used in classification trees.
}
\examples{
admission <- read.csv("http://www.ats.ucla.edu/stat/data/binary.csv")
add.factormean(factor= admission$rank, response = admission$admit)
}

