% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Vowpal_extraction7.R
\name{dt2vw}
\alias{dt2vw}
\title{Function converts data.table to Vowpal Wabbit format.}
\usage{
dt2vw(dataset, id = NULL, target = NULL, weights = NULL, char_factor,
  int_numeric)
}
\arguments{
\item{dataset}{Name of the dataset.}

\item{id}{Name of id variable column in data.}

\item{target}{Name of target variable column in data; it should be specified for
training, left out for testing}

\item{weights}{Optional vector of weight of each observation, it is a single vector,
not a part of the dataset.}

\item{char_factor,int_numeric}{Optional partion of all column names into
two groups.}
}
\description{
Internal function called by csv2vw that converts data.table to Vowpal Wabbit format.
}

